+++ rust/open/image_layer.rs
@@ -2,6 +2,8 @@ use alloc::{fmt, str::FromStr, vec::Vec};
 
 use pbf::{BitCast, ProtoRead, ProtoWrite, Protobuf};
 
+// TODO: This could be faster if we don't read in the grid data on parsing but only if the user needs it
+
 /// Track the image type
 #[derive(Debug, Clone, Copy, Default, PartialEq, Eq, PartialOrd, Ord)]
 pub enum ImageType {
@@ -20,6 +22,8 @@ pub enum ImageType {
     SVG = 5,
     /// BMP
     BMP = 6,
+    /// Unknown
+    UNKNOWN = 7,
 }
 impl BitCast for ImageType {
     fn to_u64(&self) -> u64 {
@@ -34,6 +38,7 @@ impl BitCast for ImageType {
             4 => ImageType::AVIF,
             5 => ImageType::SVG,
             6 => ImageType::BMP,
+            7 => ImageType::UNKNOWN,
             _ => panic!("unknown value: {}", value),
         }
     }
@@ -50,6 +55,7 @@ impl FromStr for ImageType {
             "AVIF" => Ok(ImageType::AVIF),
             "SVG" => Ok(ImageType::SVG),
             "BMP" => Ok(ImageType::BMP),
+            "UNKNOWN" => Ok(ImageType::UNKNOWN),
             #[tarpaulin::skip]
             _ => Err("Unknown image type"),
         }
@@ -70,6 +76,7 @@ impl fmt::Display for ImageType {
             ImageType::AVIF => "AVIF",
             ImageType::SVG => "SVG",
             ImageType::BMP => "BMP",
+            ImageType::UNKNOWN => "UNKNOWN",
         };
         write!(f, "{}", name)
     }
@@ -78,6 +85,8 @@ impl fmt::Display for ImageType {
 /// Elevation object to read from
 #[derive(Debug, Default, Clone, PartialEq)]
 pub struct ImageData {
+    /// The name of the image
+    pub name: String,
     /// The image type
     pub image_type: ImageType,
     /// The image width
@@ -89,8 +98,14 @@ pub struct ImageData {
 }
 impl ImageData {
     /// Create a new ImageData
-    pub fn new(image_type: ImageType, width: u32, height: u32, image: Vec<u8>) -> Self {
-        Self { image_type, width, height, image }
+    pub fn new(
+        name: String,
+        image_type: ImageType,
+        width: u32,
+        height: u32,
+        image: Vec<u8>,
+    ) -> Self {
+        Self { name, image_type, width, height, image }
     }
 }
 impl ProtoRead for ImageData {
@@ -100,6 +115,7 @@ impl ProtoRead for ImageData {
             1 => self.width = pb.read_varint(),
             2 => self.height = pb.read_varint(),
             3 => self.image = pb.read_bytes(),
+            4 => self.name = pb.read_string(),
             #[tarpaulin::skip]
             _ => panic!("unknown tag: {}", tag),
         }
@@ -111,5 +127,6 @@ impl ProtoWrite for ImageData {
         pb.write_varint_field(1, self.width);
         pb.write_varint_field(2, self.height);
         pb.write_bytes_field(3, &self.image);
+        pb.write_string_field(4, &self.name);
     }
 }