+++ rust/open/mod.rs
@@ -1,9 +1,9 @@
 /// Open specification for the column cache
 pub mod column_cache;
 /// Open specification for elevation storage
-pub mod elevation_data;
+pub mod grid_layer;
 /// Open specification for image storage
-pub mod image_data;
+pub mod image_layer;
 /// Open specification for the shapes
 pub mod shape;
 /// Open specification for Features
@@ -12,8 +12,8 @@ pub mod vector_feature;
 pub mod vector_layer;
 
 pub use column_cache::*;
-pub use elevation_data::*;
-pub use image_data::*;
+pub use grid_layer::*;
+pub use image_layer::*;
 pub use shape::*;
 pub use vector_feature::*;
 pub use vector_layer::*;