+++ src/open/gridLayer.ts
@@ -3,41 +3,61 @@ import { deltaDecodeArray, deltaEncodeArray } from '../util';
 
 import type { Extents } from '.';
 
-/** Elevation object to read from */
-export class ElevationData {
-  data: number[] = [];
+/** Grid object to read from */
+export class GridData {
+  #pbf: Protobuf;
+  name: string = 'default';
+  private dataPos = 0;
   extent: Extents = 8_192;
   size: number = 0;
   min: number = 0;
   max: number = 0;
 
   /**
-   * @param tag - the tag to read
-   * @param elevationData - the elevation data to mutate
-   * @param pbf - the Protobuf to pull the appropriate data from
-   * @internal
+   * @param pbf - the pbf protocol we are reading from
+   * @param end - the position to stop at
    */
-  _read(tag: number, elevationData: ElevationData, pbf: Protobuf): void {
-    if (tag === 0) elevationData.extent = pbf.readVarint() as Extents;
-    else if (tag === 1) elevationData.size = pbf.readVarint();
-    else if (tag === 2) elevationData.min = pbf.readFloat();
-    else if (tag === 3) elevationData.max = pbf.readFloat();
-    else if (tag === 4) {
-      elevationData.data = deltaDecodeArray(pbf.readPackedVarint()).map((v) =>
-        unmapValue(v, elevationData.min, elevationData.max, elevationData.extent),
-      );
-    }
+  constructor(pbf: Protobuf, end: number) {
+    this.#pbf = pbf;
+    pbf.readFields(this.#readGrid, this, end);
+  }
+
+  /**
+   * Get the raw gridded data. Reads much like a raw image.
+   * @returns - the decomposed gridded data
+   */
+  data(): number[] {
+    this.#pbf.pos = this.dataPos;
+    return deltaDecodeArray(this.#pbf.readPackedVarint()).map((v) =>
+      unmapValue(v, this.min, this.max, this.extent),
+    );
+  }
+
+  /**
+   * @param tag - the tag to read
+   * @param grid - the elevation data to mutate
+   * @param pbf - the Protobuf to pull the appropriate data from
+   */
+  #readGrid(tag: number, grid: GridData, pbf: Protobuf): void {
+    if (tag === 0) grid.extent = pbf.readVarint() as Extents;
+    else if (tag === 1) grid.size = pbf.readVarint();
+    else if (tag === 2) grid.min = pbf.readFloat();
+    else if (tag === 3) grid.max = pbf.readFloat();
+    else if (tag === 4) grid.dataPos = pbf.pos;
+    else if (tag === 5) grid.name = pbf.readString();
   }
 }
 
 /**
- * Elevation data input. It is assumed data is the actual elevation data of each point in the tile
+ * Gridded data input. It is assumed data is the actual grid data of each point in the tile
  */
-export interface ElevationInput {
-  /** The size of the square tile. The length and width are the same size */
+export interface GridInput {
+  /** The name of the grid data */
+  name: string;
+  /** The size of the square tile. The length and width MUST be the same size */
   size: number;
   /**
-   * The actual elevation data assumed to be floating point (32-bit) precise.
+   * The actual grid data assumed to be floating point (32-bit) precise.
    * It is assumed to find data you can use: `index = y * size + x`
    */
   data: number[];
@@ -49,10 +69,10 @@ export interface ElevationInput {
  * @param input - the data to encode
  * @returns - the encoded data
  */
-export function writeElevationData(input: ElevationInput): Uint8Array {
+export function writeGridData(input: GridInput): Uint8Array {
   const pbf = new Protobuf();
 
-  const { data, size, extent } = input;
+  const { name, data, size, extent } = input;
   const max = data.reduce((a, b) => Math.max(a, b), 0);
   const min = data.reduce((a, b) => Math.min(a, b), 0);
   const reMapped = data.map((v) => remapValue(v, min, max, extent));
@@ -63,6 +83,7 @@ export function writeElevationData(input: ElevationInput): Uint8Array {
   pbf.writeFloatField(2, min);
   pbf.writeFloatField(3, max);
   pbf.writePackedVarint(4, d_coded);
+  pbf.writeStringField(5, name);
 
   return pbf.commit();
 }