+++ tests/open_vector_tile/image_data.rs
@@ -16,6 +16,7 @@ mod tests {
         assert_eq!(ImageType::AVIF, "avif".into());
         assert_eq!(ImageType::SVG, "svg".into());
         assert_eq!(ImageType::BMP, "bmp".into());
+        assert_eq!(ImageType::UNKNOWN, "unknown".into());
 
         // test to string
         assert_eq!(ImageType::PNG.to_string(), "PNG");
@@ -25,11 +26,13 @@ mod tests {
         assert_eq!(ImageType::AVIF.to_string(), "AVIF");
         assert_eq!(ImageType::SVG.to_string(), "SVG");
         assert_eq!(ImageType::BMP.to_string(), "BMP");
+        assert_eq!(ImageType::UNKNOWN.to_string(), "UNKNOWN");
     }
 
     #[test]
     fn test_protobuf() {
-        let image = ImageData::new(ImageType::AVIF, 222, 333, Vec::from([1, 2, 3, 4]));
+        let image =
+            ImageData::new("test".to_string(), ImageType::AVIF, 222, 333, Vec::from([1, 2, 3, 4]));
         let mut pb = Protobuf::new();
         pb.write_message(1, &image);
 
@@ -54,6 +57,7 @@ mod tests {
         pb.write_varint(ImageType::AVIF);
         pb.write_varint(ImageType::SVG);
         pb.write_varint(ImageType::BMP);
+        pb.write_varint(ImageType::UNKNOWN);
         pb.write_varint(20);
 
         let bytes = pb.take();
@@ -66,6 +70,7 @@ mod tests {
         assert_eq!(pb.read_varint::<ImageType>(), ImageType::AVIF);
         assert_eq!(pb.read_varint::<ImageType>(), ImageType::SVG);
         assert_eq!(pb.read_varint::<ImageType>(), ImageType::BMP);
+        assert_eq!(pb.read_varint::<ImageType>(), ImageType::UNKNOWN);
 
         let result = panic::catch_unwind(AssertUnwindSafe(|| pb.read_varint::<ImageType>()));
         assert!(result.is_err());